Format convention tests
