+++ hip-more-cultural-names/common/cultures/zzz_romanian.txt
@@ -1,182 +1,182 @@
 # Romanian: ã is A-breve
 
 byzantine = {
-	romanian = {
-		male_names = {
-		    Adam_Adam
-			Adrian_Hadrianus
-			Alexandru_Alexander
-			Alin
-			Andrei_Andrew
-			Anghel_Angel
-			Aron_Aaron
-			Balc
-			Bãrbat_Barbatus
-			Basarab
-			Bogdan_Bogdan
-			Carol_Charles
-			Ciprian_Cipriano
-			Ciubar
-			Claudiu_Claudio
-			Corneliu_Cornelius
-			Corvin
-			Cosmin_Cosmas
-			Costin_Constantine
-			Cristian_Christian
-			Dan_Daniel
-			Daniel_Daniel
-			David_David
-			Dorin
-			Dragos
-			Dumitru_Demetrius
-			Fabian_Fabio
-			Florin_Floris
-			Gavril_Gabriel
-			Gelu
-			George_George
-			Gheorghe_George
-			Glad
-			Grigore_Gregory
-			Horia
-			Iacob_Jacob
-			Iancu_John
-			Ieremia_Jeremy
-			Ilas_Elio
-			Ilie_Elio
-			Ioan_John
-			Ion_John
-			Iordache_George
-			Iorgu_George
-			Iosif_Joseph
-			Iuga_George
-			Iurie_George
-			Ladislau_Vladislav
-			Lascãr
-			Latcu
-			Laurentiu_Laurence
-			Lazãr_Lazarus
-			Licã
-			Litovoi
-			Lucian_Lucius
-			Manea
-			Marin_Marinus
-			Matei_Matthew
-			Menumorut
-			Mihail_Michael
-			Mihai_Michael
-			Mihnea_Michael
-			Milos_Milosz
-			Mircea_Mircea
-			Miron
-			Moise_Moses
-			Neacsu
-			Neagu
-			Nechifor
-			Nicolaie_Nicholas
-			Nistor
-			Ovidiu_Ovidio
-			Paul_Paul
-			Petre_Peter
-			Pîrvu
-			Radovan_Radovan
-			Radu_Radu
-			Rares
-			Rãzvan
-			Remus_Remus
-			Robert_Robert
-			Roman_Roman
-			Sas
-			Sebastian_Sebastian
-			Seneslav
-			Serban
-			Sergiu_Sergios
-			Simion_Simon
-			Stan
-			Stefan_Stephen
-			Stelian
-			Stroe
-			Teodor_Theodore
-			Teodosie_Theodore
-			Tepes
-			Tibor_Tibor
-			Tihomir
-			Timotei_Timotheos
-			Toma_Thomas
-			Tudor_Theodore
-			Valentin_Valentin
-			Valeriu_Valerios
-			Vasile_Basil
-			Veaceslav_Wenzel
-			Victor_Victor
-			Vintilã
-			Vlad_Vladimir
-		}
-		female_names = {
-			Adelina_Adela
-			Adriana_Adriana
-			Afina
-			Alexandra_Alexandra
-			Alexia
-			Alina
-			Ana_Anna
-			Anastasia_Anastasia
-			Anca
-			Andra
-			Andreea_Andrea
-			Angela_Angela
-			Arina
-			Bianca_Blanche
-			Carmen_Carmen 
-			Cãtãlina
-			Clara_Clara
-			Claudia_Claudia
-			Cristina_Christina
-			Dana_Diana
-			Diana_Diana
-			Ecatarina_Catherine
-			Elena_Helen
-			Elisabeta_Elisabeth
-			Emilia_Emily
-			Felicia_Felicia
-			Floarea_Flora
-			Gabriela_Gabriella
-			Iacoba_Jakoba
-			Ioana_Joan
-			Ionela_Joan
-			Irina
-			Iulia_Juliana
-			Larisa
-			Letitia
-			Lidia_Lidia
-			Luminita
-			Mãdãlina
-			Maria_Maria
-			Mihaela
-			Monica_Monica
-			Natalia_Natalie
-			Neaga
-			Nicola_Nicholas
-			Nora
-			Olimpia_Olimpia
-			Paraschiva_Paraskeue
-			Patricia_Patricia
-			Petra_Petra
-			Raluca
-			Roxana
-			Ruxandra
-			Sabina_Sabina
-			Silvia_Silvana
-			Smârdana
-			Sofia_Sophia
-			Stana
-			Stefana_Stephania
-			Stefania_Stephania
-			Teodora Theodora
-			Tereza_Teresa
-			Vasilica
-			Voica
-			Zina
-		}
-	}
+    romanian = {
+        male_names = {
+            Adam_Adam
+            Adrian_Hadrianus
+            Alexandru_Alexander
+            Alin
+            Andrei_Andrew
+            Anghel_Angel
+            Aron_Aaron
+            Balc
+            Bãrbat_Barbatus
+            Basarab
+            Bogdan_Bogdan
+            Carol_Charles
+            Ciprian_Cipriano
+            Ciubar
+            Claudiu_Claudio
+            Corneliu_Cornelius
+            Corvin
+            Cosmin_Cosmas
+            Costin_Constantine
+            Cristian_Christian
+            Dan_Daniel
+            Daniel_Daniel
+            David_David
+            Dorin
+            Dragos
+            Dumitru_Demetrius
+            Fabian_Fabio
+            Florin_Floris
+            Gavril_Gabriel
+            Gelu
+            George_George
+            Gheorghe_George
+            Glad
+            Grigore_Gregory
+            Horia
+            Iacob_Jacob
+            Iancu_John
+            Ieremia_Jeremy
+            Ilas_Elio
+            Ilie_Elio
+            Ioan_John
+            Ion_John
+            Iordache_George
+            Iorgu_George
+            Iosif_Joseph
+            Iuga_George
+            Iurie_George
+            Ladislau_Vladislav
+            Lascãr
+            Latcu
+            Laurentiu_Laurence
+            Lazãr_Lazarus
+            Licã
+            Litovoi
+            Lucian_Lucius
+            Manea
+            Marin_Marinus
+            Matei_Matthew
+            Menumorut
+            Mihail_Michael
+            Mihai_Michael
+            Mihnea_Michael
+            Milos_Milosz
+            Mircea_Mircea
+            Miron
+            Moise_Moses
+            Neacsu
+            Neagu
+            Nechifor
+            Nicolaie_Nicholas
+            Nistor
+            Ovidiu_Ovidio
+            Paul_Paul
+            Petre_Peter
+            Pîrvu
+            Radovan_Radovan
+            Radu_Radu
+            Rares
+            Rãzvan
+            Remus_Remus
+            Robert_Robert
+            Roman_Roman
+            Sas
+            Sebastian_Sebastian
+            Seneslav
+            Serban
+            Sergiu_Sergios
+            Simion_Simon
+            Stan
+            Stefan_Stephen
+            Stelian
+            Stroe
+            Teodor_Theodore
+            Teodosie_Theodore
+            Tepes
+            Tibor_Tibor
+            Tihomir
+            Timotei_Timotheos
+            Toma_Thomas
+            Tudor_Theodore
+            Valentin_Valentin
+            Valeriu_Valerios
+            Vasile_Basil
+            Veaceslav_Wenzel
+            Victor_Victor
+            Vintilã
+            Vlad_Vladimir
+        }
+        female_names = {
+            Adelina_Adela
+            Adriana_Adriana
+            Afina
+            Alexandra_Alexandra
+            Alexia
+            Alina
+            Ana_Anna
+            Anastasia_Anastasia
+            Anca
+            Andra
+            Andreea_Andrea
+            Angela_Angela
+            Arina
+            Bianca_Blanche
+            Carmen_Carmen 
+            Cãtãlina
+            Clara_Clara
+            Claudia_Claudia
+            Cristina_Christina
+            Dana_Diana
+            Diana_Diana
+            Ecatarina_Catherine
+            Elena_Helen
+            Elisabeta_Elisabeth
+            Emilia_Emily
+            Felicia_Felicia
+            Floarea_Flora
+            Gabriela_Gabriella
+            Iacoba_Jakoba
+            Ioana_Joan
+            Ionela_Joan
+            Irina
+            Iulia_Juliana
+            Larisa
+            Letitia
+            Lidia_Lidia
+            Luminita
+            Mãdãlina
+            Maria_Maria
+            Mihaela
+            Monica_Monica
+            Natalia_Natalie
+            Neaga
+            Nicola_Nicholas
+            Nora
+            Olimpia_Olimpia
+            Paraschiva_Paraskeue
+            Patricia_Patricia
+            Petra_Petra
+            Raluca
+            Roxana
+            Ruxandra
+            Sabina_Sabina
+            Silvia_Silvana
+            Smârdana
+            Sofia_Sophia
+            Stana
+            Stefana_Stephania
+            Stefania_Stephania
+            Teodora Theodora
+            Tereza_Teresa
+            Vasilica
+            Voica
+            Zina
+        }
+    }
 }
